9.5. Финальный отсчёт

+++ src/entities/film/controller/film-controller.ts
@@ -12,20 +12,23 @@ import UpdateFilmDto from '../dto/film-update-dto.js';
 import HttpError from '../../../common/errors/http-error.js';
 import CreateFilmDto from '../dto/film-create-dto.js';
 import {MAX_FILMS_COUNT} from '../films-constants.js';
-import SearchFilmsResponse from '../response/SearchFilmsResponse.js';
 import {ICommentService} from '../../comment/comments-service-interface.js';
 import {ValidateDtoMiddleware} from '../../../common/middlewares/validate-dto-middleware.js';
 import {DocumentExistsMiddleware} from '../../../common/middlewares/document-exist-middleware.js';
 import {ValidateObjectIdMiddleware} from '../../../common/middlewares/validate-object-id-middleware.js';
 import CommentResponse from '../../comment/response/comment-response.js';
 import {PrivateRouteMiddleware} from '../../../common/middlewares/private-route-moddleware.js';
+import {SearchFilmsParams} from '../dto/search-films-params.js';
+import FilmSearchItem from '../response/film-search-item.js';
+import {IUserService} from '../../user/service/user-service-interface.js';
 
 @injectable()
 export default class FilmController extends Controller {
   constructor(
     @inject(Component.ILogger) logger: ILogger,
     @inject(Component.IFilmService) private readonly filmsService: IFilmService,
-    @inject(Component.ICommentService) private readonly commentService: ICommentService
+    @inject(Component.ICommentService) private readonly commentService: ICommentService,
+    @inject(Component.IUserService) private readonly userService: IUserService
   ) {
     super(logger);
 
@@ -34,35 +37,59 @@ export default class FilmController extends Controller {
     this.addRoute({
       path: '/',
       method: HttpMethod.Get,
-      handler: this.index
+      handler: this.search
     });
     this.addRoute({
-      path: '/create',
+      path: '/',
       method: HttpMethod.Post,
       handler: this.createFilm,
       middlewares: [
         new PrivateRouteMiddleware(),
         new ValidateDtoMiddleware(CreateFilmDto)]
     });
+
+    this.addRoute({
+      path: '/watch',
+      method: HttpMethod.Post,
+      handler: this.postToWatch,
+      middlewares: [new PrivateRouteMiddleware()]
+    });
+    this.addRoute({
+      path: '/watch',
+      method: HttpMethod.Delete,
+      handler: this.deleteToWatch,
+      middlewares: [new PrivateRouteMiddleware()]
+    });
+
+    this.addRoute({
+      path: '/watch',
+      method: HttpMethod.Get,
+      handler: this.getToWatch,
+      middlewares: [new PrivateRouteMiddleware()]
+    });
+
+    this.addRoute({
+      path: '/promo',
+      method: HttpMethod.Get,
+      handler: this.getPromo,
+    });
     this.addRoute({
       path: '/:id',
       method: HttpMethod.Get,
       handler: this.get,
       middlewares: [
-        new ValidateObjectIdMiddleware('id'),
-        new ValidateDtoMiddleware(UpdateFilmDto)
+        new ValidateObjectIdMiddleware('id')
       ]
     });
     this.addRoute({
-      path: '/:id',
+      path: '/',
       method: HttpMethod.Patch,
       handler: this.update,
       middlewares: [
         new PrivateRouteMiddleware(),
-        new ValidateObjectIdMiddleware('id'),
         new ValidateDtoMiddleware(UpdateFilmDto),
         new DocumentExistsMiddleware(this.filmsService, 'Film', 'id')
-        ]
+      ]
     });
     this.addRoute({
       path: '/:id',
@@ -75,16 +102,6 @@ export default class FilmController extends Controller {
       ]
     });
     this.addRoute({
-      path: '/search',
-      method: HttpMethod.Get,
-      handler: this.search,
-    });
-    this.addRoute({
-      path: '/promo',
-      method: HttpMethod.Get,
-      handler: this.getPromo,
-    });
-    this.addRoute({
       path: '/:filmId/comments',
       method: HttpMethod.Get,
       handler: this.getComments,
@@ -95,15 +112,9 @@ export default class FilmController extends Controller {
     });
   }
 
-  public async index(_req: Request, res: Response): Promise<void> {
-    const films = await this.filmsService.find();
-    const filmsResponse = fillDTO(FilmResponse, films);
-    this.ok(res, filmsResponse);
-  }
-
   public async createFilm(req: Request<Record<string, unknown>, Record<string, unknown>, CreateFilmDto>,
     res: Response): Promise<void> {
-    const {body} = req;
+    const {body, user } = req;
 
     const existFilm = await this.filmsService.findByName(body.title);
 
@@ -115,8 +126,8 @@ export default class FilmController extends Controller {
       );
     }
 
-    const result = await this.filmsService.create(body);
-    this.created(res, fillDTO(FilmResponse, result))
+    const result = await this.filmsService.create({...body, user: user.id});
+    this.created(res, fillDTO(FilmResponse, result));
   }
 
   public async update(
@@ -132,10 +143,15 @@ export default class FilmController extends Controller {
         'FilmController'
       );
     }
-
-    console.log({...body, userId: user.id})
-    const result = await this.filmsService.update({...body, userId: user.id});
-    this.ok(res, fillDTO(FilmResponse, result))
+    if (existFilm.user?.id !== user.id) {
+      throw new HttpError(
+        StatusCodes.FORBIDDEN,
+        'Only the owner can edit films. ',
+        'FilmController'
+      );
+    }
+    const result = await this.filmsService.update({...body, user: user.id});
+    this.ok(res, fillDTO(FilmResponse, result));
   }
 
   public async get(
@@ -161,7 +177,7 @@ export default class FilmController extends Controller {
     const {params, user} = req;
     const existFilm = await this.filmsService.findById(params.id);
 
-    if (existFilm?.userId?.id !== user.id) {
+    if (!existFilm) {
       throw new HttpError(
         StatusCodes.UNPROCESSABLE_ENTITY,
         `Film with id «${params.id}» not exists.`,
@@ -169,24 +185,36 @@ export default class FilmController extends Controller {
       );
     }
 
+    if (existFilm.user?.id !== user.id) {
+      throw new HttpError(
+        StatusCodes.FORBIDDEN,
+        'Only the owner can edit films. ',
+        'FilmController'
+      );
+    }
+
     await this.filmsService.deleteById(params.id);
+    await this.commentService.deleteByFilmId(params.id);
     this.ok(res, true);
   }
 
   public async search(
-    {params}: Request<Record<string, string>>,
+    req: Request<unknown, unknown, unknown, SearchFilmsParams>,
     res: Response): Promise<void> {
-    let limit = params.limit;
-    if (!limit) {
-      limit = MAX_FILMS_COUNT.toString();
+    const { limit, genre } = req.query;
+    const numberLimit = limit ?? MAX_FILMS_COUNT;
+    let searchResult;
+    if (!genre) {
+      searchResult = await this.filmsService.findByLimit(numberLimit);
     }
-    let numberLimit: number = Number(limit);
-    const searchResult = await this.filmsService.findByGenre(params.genre, numberLimit);
+    else {
+      searchResult = await this.filmsService.findByGenre(genre, numberLimit);
+    }
 
     this.send(
       res,
       StatusCodes.OK,
-      fillDTO(SearchFilmsResponse, searchResult)
+      fillDTO(FilmSearchItem, searchResult)
     );
   }
 
@@ -204,5 +232,23 @@ export default class FilmController extends Controller {
     res: Response): Promise<void> {
     const comments = await this.commentService.findByFilmId(params.filmId);
     this.ok(res, fillDTO(CommentResponse, comments));
+  }
+
+  async getToWatch(req: Request<Record<string, unknown>, Record<string, unknown>>, _res: Response): Promise<void> {
+    const {user} = req;
+    const result = await this.userService.findToWatch(user.id);
+    this.ok(_res, fillDTO(FilmSearchItem, result));
+  }
+
+  async postToWatch(req: Request<Record<string, unknown>, Record<string, unknown>, { filmId: string }>, _res: Response): Promise<void> {
+    const {body, user} = req;
+    await this.userService.addToWatch(body.filmId, user.id);
+    this.noContent(_res, {message: 'Успешно. Фильм добавлен в список "К просмотру".'});
+  }
+
+  async deleteToWatch(req: Request<Record<string, unknown>, Record<string, unknown>, { filmId: string }>, _res: Response): Promise<void> {
+    const {body, user} = req;
+    await this.userService.deleteToWatch(body.filmId, user.id);
+    this.noContent(_res, {message: 'Успешно. Фильм удален из списка "К просмотру".'});
   }
 }
